# internal/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateRoleRequest:
    properties:
      description:
        maxLength: 200
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  dto.RoleResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.UpdateRolePermissionsRequest:
    properties:
      permission_ids:
        items:
          type: string
        type: array
    required:
    - permission_ids
    type: object
  dto.UpdateRoleRequest:
    properties:
      description:
        maxLength: 200
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      errors:
        items:
          $ref: '#/definitions/response.ErrorDetail'
        type: array
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
      source: {}
    type: object
  response.Meta:
    properties:
      count:
        type: integer
      current_page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Server for Load Stuffing Calculator application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Load Stuffing Calculator API
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of roles.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.RoleResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Creates a new role. Requires admin privileges.
      parameters:
      - description: Role Creation Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new role
      tags:
      - roles
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a role by ID. Requires admin privileges.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a role
      tags:
      - roles
    get:
      consumes:
      - application/json
      description: Retrieves role details by ID.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get a role by ID
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Updates an existing role. Requires admin privileges.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Role Update Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a role
      tags:
      - roles
  /roles/{id}/permissions:
    get:
      consumes:
      - application/json
      description: Retrieves a list of permission IDs assigned to a role.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get permissions for a role
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Replaces all permissions for a role. Requires admin privileges.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Role Permissions Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRolePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update permissions for a role
      tags:
      - roles
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"